backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
pytesseract==0.3.10
Pillow==10.4.0
python-dotenv==1.0.1
numpy==2.1.0